Updated base request class + documentation
A method was added to the base request class that passes on the given
parameters into requests.get(). It does so because now the only file
that must import this method is the file with the base requests. Not
onlt that, but it returns the response.text, so that each child method
does not have to get the text from the dictonary.
Documentation had minor changes, the main change is that lines of
documentation stating the request type = get have been removed. I did
not realise all of them were get, so there is no variation. Without
variation, there is no need to state the request type for each method.