Progress on main, minor changes elsewhere
The first steps were made in the main.py file. While doing so, I had to
make minor changes to some of the api requests files. Two of the methods
were changing the base url shared by all of the requests methods. That
is now fixed. The main file is not yet properly documented and also it
is a bit sloppy at the moment. This will be changed shortly.